[3/b] Pattern match GitHub issue inputs (#31422)
## Summary

If the user inputs a string which is not a valid git branch or if the user inputs a valid GitHub issue, defer generating PR title and branch name to Claude Code. Pattern match special input types (e.g. GitHub issue URLs) to glom on extra tools and extra context to the Claude Code invocation.