minimal asset materialization health class (#31115)
## Summary & Motivation
Adds a smaller class that we can use to derive the materialization
health state for an asset. This class just stores the number of failed
partitions and currently materialized partitions for each asset (this
treats non-partitioned assets as if they are a partitioned asset of 1
partition). We can compute the health status for an asset from this
information, which allows us to bypass deserializing potentially large
EntitySubsets

Also adds a `last_materialization_timestamp` attr to
`AssetMaterializationHealth` and the smaller class.