[ui] Fix missing loading bar on the RunsFeedTable
It looks like the Runs feed table has not been showing it’s indeterminate loading bar, and Daniel flagged that it was confusing when switching tabs, since there’s a delay before the rows reflect the new tab.

This is happening because the IndeterminateLoadingBar is rendered inside the `overflow:hidden` portion of the run feed table. Both code paths inside `content()` render the loading bar, so I moved it up one layer so that it is rendered in the same place but outside the overflow: hidden container.