[ui] Prevent hidden chart tooltips from blocking clicks (#31564)
## Summary & Motivation

This is an interesting one we noticed during testing of the Billing page. When the chart tooltip is closed, it remains where it was in the DOM and blocks clicks on elements beneath it. Adding `display:none` changes the radix popover container size to 0px and fixes the issue.

This could have an impact if we had tooltip animations on the opacity, but it appears we do not.

<img width="1171" height="680" alt="image" src="https://github.com/user-attachments/assets/546fff4a-e0e3-40fb-ae01-20f111bfd466" />