[ui] Don’t show triangle on single asset lineage sidebar repo (#31585)
## Summary & Motivation

https://linear.app/dagster-labs/issue/FE-815/dont-show-collapse-icon-if-theres-only-1-code-location

If there is only one code location, but more than one asset group, the
asset graph sidebar forces the code location open all the time. This PR
hides the disclosure triangle in this case rather than it just being
broken.

Sidenote: We don't have stories for this component and this behavior is
tricky - to repro this I had to adjust the toys repo, couldn't see it in
elementl prod or dogfood/hooli. It's also interesting that if there is
just one code location and one asset group we get rid of all the
hierarchy and just show the assets, but we don't get rid of this code
location row if there's just one code location. Will come back and add
stories in a follow-up.

<img width="358" height="442" alt="image"
src="https://github.com/user-attachments/assets/e10a7848-d8f5-4a74-ac65-e495d468c47d"
/>

Co-authored-by: bengotow <bgotow@elementl.com>