Cancel all cancellable runs in a backfill iteration instead of some of them
Summary:
We currently cancel runs in batches of 50 on each iteration. This made more sense in a pre-threadpool world, but now it just artifically makes it slower to cancel runs when there are many to cancel.

Also pick runs to cancel in ascending order rather than descending, to ensure that we get the runs that are most likely to be launched first.

Test Plan: BK, simulate cancellation of a large backfill locally and watch logs