Introduce ToxFactor
I want to start threading --split through our PackageSpec so I can
parallelize pytest tests a bit more. And I sometimes want different
splits for different tox factors.

So this introduces a new data class that I can later extend to include
that config. Expectation is that this particular change is a no-op for
the build.