Parallelize the build
Looking at a recent waterfall chart on Buildkite:

https://buildkite.com/dagster/dagster-dagster/builds/130942/waterfall

Our longest running non-parallelizable step is pyright:

https://buildkite.com/organizations/dagster/analytics/suites/dagster-steps/tests/ff1d16cc-3650-85a9-8576-d01b4701f4e1?period=7days

It takes about 15 minutes. Which means that's currently our limiting
factor for build speed.

All other longer steps are pytest steps that in theory can be
parallelized.

So this just goes through the list and eyeballs anything that takes
longer than ~15 minutes and tries to divide them in two.

And we'll use Buildkite to find out if there are any issues with that
parallelization (for example, inappropriately coupled tests that rely on
each other).