fix issue where an asset was declared a stub if it was a stub in any code location
Summary:
Unlike these other properties, an asset should only be called a stub if it is a stub in the actual asset definition that we picked.

Load an asset job with the 'filter stubs' toggle on when it defined in another code location as a stub, the toggle no longer excludes it.