Update BaseFallthroughRoot to handle no-job schedules and automations
Right now, if you have no jobs, but have schedules or automation conditions
to materialize assets, the default dagster landing page redirects to the first
asset group.

This proposes to change instead to redirect to /overview (like we would if
there were jobs), but to use the automation grouping instead, since that
should list the materializations in a reasonable fashion.