Update type of use_partition_load_context (#31644)
## Summary & Motivation

This fixes a few issues with this decorator.

The main one was that it erased all type information of the decorated
function. So basically all functions decorated with it became
unannotated, which was bad.

The secondary one was that it was possible to annotate a function on a
class that did not have a `_partition_loading_context` property on it.
Now you will get type errors if you try to do that.

It's a bit cursed but hey it works.

note: added two "new" `# type: ignore`s in the test code, but those
actually were removed with the change that added this decorator (because
the decorator made the type of the things Any). so this is just
restoring the pre-decorator behavior there.

## How I Tested These Changes

## Changelog

NOCHANGELOG