Always depend on tzdata, not just on Windows
Summary:
The python:slim packages seem to have upgraded to a new version of debian in the last 24 hours that doesn't have all the same timezone strings that the previous version did, which means that declaring timezones like US/Central can start failing without having this dependency installed. Add it as an explicit dependency to mitigate that (before, we were only adding it on Wdinows).