Always depend on tzdata, not just on Windows (#31741)
## Summary & Motivation
The python:slim packages seem to have upgraded to a new version of
debian in the last 24 hours that doesn't have all the same timezone
strings that the previous version did, which means that declaring
timezones like US/Central can start failing without having this
dependency installed. Add it as an explicit dependency to mitigate that
(before, we were only adding it on Wdinows).


## How I Tested These Changes
BK

## Changelog
Added `tzdata` as a dependency to `dagster`, to ensure that declaring
timezones like `US/Central` work in all environments.