[backfills] allow run config in PartitionBackfill (#31229)
## Summary & Motivation

This is a hotly-requested feature. This PR allows runConfig to be set when creating a PartitionBackfill, and has that run config set on all runs for that backfill. This is similar to how we handle backfill tags.

I was originally concerned that there would be complications with supplying too much config to these runs (not all runs will contain all ops), but it turns out that our config system handles that without issue, as seen in the added tests.

Also adds this to the graphql schema, but does not update the UI to allow for this.

## How I Tested These Changes

## Changelog

NOCHANGELOG