Run test_docs_component_type_success in isolated dir
I suspect this is failing on Buildkite because of how we mount /tmp
to the running container although I don't quite know why that matters.
Potentially collision between multiple tests running using the same
/tmp/.venv?

https://buildkite.com/dagster/dagster-dagster/builds/132825#0198c330-fb36-4ec7-a6f6-ddd4f5f708ba/474-540

Regardless, I'd like to see if this works. The test has always succeeded
locally but reliably fails remotely (it's just been muted for a while).