Mark k8s garbage collection tests as flaky
Another example of something that universal retries was papering over.
These seem to time out a lot and succeed on retry. It's currently the
least reliable test in our test suite.

https://buildkite.com/organizations/dagster/analytics/suites/dagster/tests/ab6efe0a-2745-8d82-859c-e252888eaee1

We've already bumped the timeout:

https://github.com/dagster-io/dagster/pull/31336

We can keep bumping if we'd like. This just adds a couple of retries to
continue papering over but ultimately, we'll want to decide if this test
is providing value and if so, dig deeper on why it's so unreliable.