Hard skip test_threaded_submit_backfill (#31880)
I suspect we're deadlocking from time to time. When it happens, it
hanges the entire buildkite build and there's no way to recover.

This has been skipped in Buildkite for a while with a backlogged task to
dig deeper - I'm just moving the virtual skip to exist in code instead.