[2/n] thread StateVersionMapping through DefinitionsLoadContext, initial StateBackedComponent impl (#31696)
## Summary & Motivation

The main thrust of this PR is to thread the StateVersions information through the DefinitionsLoadContext. There are two modes to loading definitions, INITIALIZATION and RECONSTRUCTION.

Before this PR, any time you were in INITIALIZATION mode, this meant all data would be recalculated from scratch from external systems. Now, we assume that even in this mode, the DefinitionsLoadContext will still be aware of the versions of any state objects that should be used for loading. In most cases, this version information will just be the latest set of versions that are available, but in the context of rollbacks, this may be set explicitly to some earlier value. We pass this information by setting it directly on the DefinitionsLoadContext when it is initially set. We also make sure that we set the version inforation in the `pending_reconstruction_metadata`, so that we can pull from it in RECONSTRUCTION mode.

In RECONSTRUCTION mode (as mentioned above), the state version information is always available in the reconstruction metadata and can be fetched from there. For anything querying the DefinitionsLoadContext, they can fetch the state version of a given definitions key with a special context manager on the DefinitionsLoadContext which handles downloading the file and tracking the right place to pull the state information from.

The CodeLocationDeployData is modified in this PR to contain the state version information that was used. Conceptually, state versioning info is very similar to container image -- the ideal situation is one in which the combination of (container_image, state_versions) fully describes the set of definitions in a code location -- of course there's nothing stopping anyone from writing non-deterministic code, but this is a big step closer to avoiding redploys being ND.

NOTE: it's bad form for this PR to also include the initial StateBackedComponent class, but it's quite hard to write a test without this in there.

## How I Tested These Changes

## Changelog

NOCHANGELOG