[5/n] Display version info in the ui (#31784)
## Summary & Motivation

We want to be able to visualize the state versions that are being used in a given code location alongside the latest available ones

This UI only appears in cases where you're using state-backed components, so this should not impact existing users at all.

Normal:

![Screenshot 2025-08-19 at 1.40.11 PM.png](https://app.graphite.dev/user-attachments/assets/8b846037-eb77-4704-bc80-5d34bc5db36c.png)

Out of date:
![Screenshot 2025-08-19 at 1.41.00 PM.png](https://app.graphite.dev/user-attachments/assets/ca62f4dd-97bd-4277-a69b-db19e68f78b4.png)

There are many other places where we might want to thread this information into in the future, but for now this serves as more of a useful diagnostic thing as we start experimenting with this capability.

## How I Tested These Changes

## Changelog

NOCHANGELOG