[6/n] Add serialized_defs_state_info to CodeLocationDeployData (#31713)
## Summary & Motivation

This adds information about the defs_state that was used to deploy a Code Location to the CodeLocationDeployData.

It's stored as a pre-serialized string (rather than the actual DefsStateVersions object) for simplicity -- this way, we don't need to move the DefsStateVersion object into dagster_shared in order for the dagster-cloud-cli package to access it. Also I could imagine this avoiding some weird version incompatibility stuff when we round-trip the object, although that's mostly secondary.

## How I Tested These Changes

## Changelog

NOCHANGELOG