[7/n] Update get_defs_state_key functionality (#31861)
## Summary & Motivation

We want this to be based on something more stable than the path to the component. This uses the class name, which feels like a fairly-straightforward option.

This PR adds in error-checking functionality, providing a useful error if/when multiple components are detected with the same defs state key.

Also fixes an issue w/ the ComponentTreeDependencyTracker where a single instance was shared across all ComponentTrees (and renames it to be more general)

## How I Tested These Changes

## Changelog

NOCHANGELOG