[dagster-dbt] Update DbtProjectComponent with configurable CLI args (#31881)
## Summary & Motivation

This updates the DbtProjectCompnent to allow users to pass in custom arguments to the cli invocation.

We could definitely get away with the signature of the `cli_args` field being just `list[str]`, but the way the `--vars` parameter works means that it'd be pretty tough to actually format things in a yaml document (because that param expects json-formatted data). So I added the option to just specify things in dictionary format.

In theory, this is technically a bit overly-specified, in that it's possible that a user's component subclass does not use this configuration at all (i.e. maybe it invokes two completely separate commands or something). However, in the vast majority of cases, it seems likely that people will only invoke one dbt command for the actual execution of their assets (maybe they do other commands for other reasons afterwards), and so it feels ok to name this field this way.

## How I Tested These Changes

## Changelog

The DbtProjectComponent now takes an optional `cli_args` configuration to allow customizing the command that is run when your assets are executed.