[ui] Insights v2: Allow clicking line chart for prev data only (#31910)
## Summary & Motivation

For Insights v2, allow clicking on line chart segments that have only previous data and not current data, e.g. for "This week" and "This month".

Also updated the rows of the tooltip to show the actual dates, instead of "current period" and "previous period", since it wasn't clear before what the previous period was.

<img width="404" height="285" alt="Screenshot 2025-08-22 at 14 32 59" src="https://github.com/user-attachments/assets/c70c0a58-630f-4cb9-a337-f34ce3f0ed70" />
<img width="411" height="255" alt="Screenshot 2025-08-22 at 14 33 43" src="https://github.com/user-attachments/assets/5f887d68-977d-4a7b-bfa4-300bfb0ff853" />


## How I Tested These Changes

Proxy, view charts and hover/click to verify correct behavior.