Use a repo-scoped asset graphh in changedReasons when it is called on a repo-scoped node (#31903)
Summary:
changedReasons is sometimes called with workspace-scoped asset nodes
(often on a smaller set of asset nodes) and sometimes with repo-scoped
asset nodes. In the former case
https://github.com/dagster-io/dagster/pull/26904 helps a lot, but in the
latter case it does not help, since a) we already have the full
repo-scoped asset graph in memory anyway, for the branch at least, and
b) its often called on a bunch of asset keys at once, and the large
number of point lookups on the workspace asset graph actually hurt
rather than help.

So instead do a partial revert of that PR and only use the
workspace-scoped asset graph when you're in a workspace scoped asset
graph.