[backfill-config] Fix asset selection (#31938)
## Summary & Motivation

Fixes an issue where if you have a sub-selection of assets, in your backfill, we previously would require config to be provided for all assets in the code locaiton.

The skipConfigQuery pattern is a little suspect but it was the simplest solution I could find.

Also makes it possible to click out of the config dialog without saving, as it's kinda annoying not to be able to.

## How I Tested These Changes

manually

## Changelog

Fixed an issue that would require config provided to backfills to contain config for all assets in the conde location rather than just the selected ones.