Allow for partial ComponentTreeReloading (#31866)
## Summary & Motivation

**NOTE:** this implementation is pretty jank, and mostly a proof of concept / addition of a test case prior to the next PR in the stack, which formalizes this caching concept a bit more. Arguably, I should just merge these prs together, but I do a bunch of other refactoring in the upstack PR as well, so I didn't want the core logical bit to be tied into that. 

this reworks https://github.com/dagster-io/dagster/pull/31482 to have a more explicit concept of a "cache key" that is wired through any cached methods / properties. Each path has an associated cache key, allowing us to update this cache key out of band when certain events happen.

In this pr, the cache key will only ever get invalidated manually, but upstack, we'll add functionality to invalidate the cache key when the state of a state-backed component is updated. this will make it possible to do a partial reload of only the impacted components.

## How I Tested These Changes

## Changelog

NOCHANGELOG