[ui] Avoid unmount + remount of viewport when hiding explorer sidebar (#31956)
## Summary & Motivation

Related:
https://dagsterlabs.slack.com/archives/C03CCE471E0/p1755645402070769

Right now, when you close the sidebar on the asset lineage page the
component hierarchy changes so the SVGViewport is unmounted and
remounted. We use a context to preserve some viewport state within the
SVGViewport between layout changes, but this is higher in the tree.

We support passing a null `second` view to the SplitPaneContainer and
animates the second panel in when the value changes. This PR just allows
us to do the same for the `first` pane so we can hide the sidebar panel
without remounting the viewport.

Co-authored-by: bengotow <bgotow@elementl.com>