Optimize set operations on two DefaultPartitionSubsets and is_empty on DefaultPartitionsSubset
Summary:
The base implementation creates a new copy of the underlying set and uses with_partition_keys which can create new copies and make a real difference for large subsets. Manipulate the underlying sets directly when possible.

BK, microbenchmark script that takes two DefaultPartitionSubsets and subtracts them:

```
from dagster._core.definitions.partitions.subset.default import DefaultPartitionsSubset

seta = DefaultPartitionsSubset({str(a) for a in range(100000)})

setb = DefaultPartitionsSubset({str(a) * 2 for a in range(100000)})

for i in range(200):
    print(str(i))
    setc = seta - setb
```

> Insert changelog entry or delete this section.