[dagster-dbt] Fix test (#31988)
## Summary & Motivation

The original implementation would work in the real world, but not in tests, so this goes with a more general approach.

Side note: I very much hate the pattern of "property that might just throw an exception if you access it" instead of "property that might just be None", but it is what it is I guess.

## How I Tested These Changes

## Changelog

NOCHANGELOG