[ui] Run compute log: Add title to step selector items (#31980)
## Summary & Motivation

On the logs toolbar step selector, add a "title" to allow displaying the entire string on hover. This is useful in cases where step names are super long.

I wasn't able to get `data-tooltip` to work here, possibly because of the DOM nesting involved in the select menu. But since it would require hovering anyway, it works to just show the native title.

## How I Tested These Changes

Hover on items in the step selector dropdown in the compute log UI, verify that the title appears with the full string.