[buildkite] re-enable automatic retries (#31996)
The set of exit status codes that are targeted in this configuration aim
to apply automatic retries only to ephemral infrastructural issues.
Notably exit code `1` which is what happens on normal test failures is
not inlucded here.

I have removed the entry for exit code `2` which happen in makefile
based bk steps. I think these failures triggering retries was the
motivation behind #31768.

If we find that infastructural issues that we believe we want automatic
retries on are surfacing as `2` we can update the target make commands
to explicitly return `1` on non-transitive failures

## How I Tested These Changes

eyes