[dg] load_from_defs_folder project_root -> path_within_project (#31992)
When `src` based layouts are built for distribution, the `src` directory
is removed so relative pathing all the way to the directory will fail as
the path changes between environments.

The underyling functionality here is flexible and looks upwards of the
provided directory to find the config so embrace that and rename the
argument accordingly.

## How I Tested These Changes

existing coverage