[dg] make dg projects work when properly installed (#31991)
dg projects currently do not work when installed non-editable as we seek
to read a single source of truth of config from pyproject.toml / dg.toml
at runtime which is not included in the package output by default.

To address this, update hatch settings to include the file in output.

## How I Tested These Changes

added test