[asset graph] Minor sidebar refactor + Fix asset graph sidebar tree view auto-scrolling (#32006)
## Summary & Motivation

Currently when you click on an asset there is a chance you get scrolled
to another instance of the same asset elsewhere in the tree. This is
because the current onClick handlers on record the key of the asset
being clicked and not the exact path. To fix this record the exact path
of the node that is clicked in the sidebar.

## How I Tested These Changes


https://github.com/user-attachments/assets/d1207215-aac0-4f4c-bd82-a5bd9be75655