Return an empty list in assetsOrError if an empty list is passed in (#32025)
## Summary & Motivation
Before this would return every asset when an empty list was passed in.
Now it returns no assets.

## How I Tested These Changes
New test cases

## Changelog
Fixed an issue where passing in an empty list in the `assetKeys`
argument of the `assetsOrError` field in the GraphQL API would return
every asset instead of an empty list of assets.