[dagster-omni] initial component implementation (#31775)
## Summary & Motivation

This adds an OmniComponent to the (previously empty) dagster-omni library.

Apologies in advance as it's a big PR, but figured it's useful to have all the end-to-end stuff in here so you can see the full vision.

This is the first "real" implementation of a StateBackedComponent, so the general patterns here are new as well.

Some things to note:

- Still using the `XWorkspace` abstraction to separate out configuration for linking to / configuring how we query information from a given external API. I found this to be useful separation of concerns, and you can imagine future worlds in which we have more generic workspace abstractions that users can subclass from
- NOT using a translator abstraction. I think this should be standard practice for net-new integrations. Translation is really just a single function and creating a whole object for that purpose is very heavyweight. The component can handle this, and we can configure that function using the `translation` method used by other integrations without having to construct a new class.
- NOT using resources. Because there is no spinup cost for instantiating an OmniWorkspace (it just holds values until it's used, it doesn't immediately spin up a session or anything like that), there is no reason to put it into the fairly-complicated resource lifecycle.

In terms of state-backed things, the general pattern here (which I believe makes sense for most integrations of this sort) is:

1. Create a mapping from the raw API responses to structured `@record` objects that we can serialize.
2. Create methods to fire off queries and convert into this serializable format
3. Create a single top-level `@record` object for `XState`, which contains all of the sub-objects we get from the API

This step of transforming the raw API responses into structured objects is (imo) necessary because it enforces schema validation before writing, and formalizes the shape of responses we expect. Also means we can handle reading old state data with new dagster code gracefully (with serdes magic)

## How I Tested These Changes

## Changelog

NOCHANGELOG