[Asset health] Only fetch health for assets that have definitions (#32044)
## Summary & Motivation
Fetching health for assets without definitions causes expensive stuff to
happen on the backend so lets avoid sending health queries for them.

A previous version of this did the filtering in the query function but
this pr moves it to the hook itself to avoid batching/chunking the keys
that dont have definitions.

## How I Tested These Changes

loaded a bunch of orgs included ones that previously caused issues.