[1/so] Represent null state versions (#31984)
## Summary & Motivation

This set of changes allows us to represent null state versions, which can happen when a compnent attempts to access state while loading but no state is available.

This is important for showing things in the UI for code locations that have never generated state for certain components, and alerting users to that fact.

In order for this to work, we explicitly update the defs_state_info at runtime for cases where it is null.

## How I Tested These Changes

## Changelog

NOCHANGELOG