[2/so] Add `dg plus deploy refresh-defs-state` command (#31985)
## Summary & Motivation

This adds a new optional `dg plus deploy refresh-defs-state` command. In order for this to work, we need to bring in the GraphQLDefsStateStorage and create a version of it that can be instantiated with just the available information in the DagsterPlusCliContext object. All this object needs to do is fire off graphql / rest queries, so things work out nicely as this just requires a user token, a url, and a deployment name, all of which we have available.

High level changes:

- factored out common utilities from the existing `dg utils refresh-defs-state` command (which i renamed from `dg utils refresh-component-state` as I think the new things sounds nicer)
- made `DefsStateInfo` and `DefsKeyStateInfo` DagsterModels so that they can be sub-objects in a pydantic schema. While this does have theoretical performance implications, the total number of StateBackedComponents in a single project is pretty well-bounded (e.g. we would never have 10k of these), and it means we can get rid of the custom logic in the from_dict/to_dict bit and rely on pydantic for all of that, which is nice
  - note: i moved this into `dagster_shared.serdes.objects.models` to avoid hitting the expensive `pydantic` import every time we import `dagster_shared.serdes.objects` for other stuff
- moved the core definition of the GraphQLDefsStateStorage into this repo so that it can be subclassed for use in the CLI
- made it so that we put defs_state_info into the code location deploy document during the build process

## How I Tested These Changes

Honestly, writing a test was pretty challenging as we need to mock out most of the interesting stuff. Open to ideas on more thorough testing here but I was able to at least confirm that we do actually invoke the relevant `StateBackedComponent.refresh_state()` commands.

## Changelog

NOCHANGELOG