[components] Error if attributes are passed to a component without a model (#31459)
## Summary & Motivation

Ran into this in the wild, and was very confused. The net result is that before this change, we would silently drop the `attributes` dictionary and then error when constructing the component class. Now we can error in the yaml itself.

## How I Tested These Changes

## Changelog

Fixed an issue that could cause confusing errors when attempting to supply `attributes` configuration to `Component` subclasses that did not inherit from `Resolvable`.