[sensor-preview] Fix bug where run_keys were not respected (#31452)
## Summary & Motivation

This matches the behavior of regular sensors in which we filter out run requests that have a run_key matching an existing run.

These runs do not even show up in the UI, so a similar behavior here seems acceptable

## How I Tested These Changes

unit, manual

## Changelog

Fixed a bug with the sensor preview behavior that would cause run requests contianing `run_key`s that had already been submitted to show up in the preview result.