[dbt-cloud] Emit asset events before step failure (#32035)
## Summary & Motivation

Previously, the `poll_run` command would immediately raise an error when
a run failure was detected. This is a bit confusing (in my opinion), and
is not actually noted in the docstring. I've updated the code to:

- Not error at all in `poll_run`, instead add a new `wait()` method on
the run_handler
- removed `wait_for_success` and instead replaced that call with a
`wait()`, followed by a `raise_for_status`.
- Add a `raise_for_status` method on the run object. this is modeled
after the `responses` library, where you get a `Response` object and
then can decide if/when you want to generate an error based off of that
response.
- Removed a check for the test status being PASS when emitting check
events -- this brings it in line with the existing implementation for
dbt core

With those changes, I was able to move the error raising behavior to the
bottom of the `cli()` command, meaning we always yield asset events
BEFORE the failure is raised.

## How I Tested These Changes

## Changelog

[dbt-cloud] Fixed an issue that would cause the DbtCloudWorkspace to
error before yielding asset events if the associated dbt Cloud run
failed. Now, it will raise the error *after* all relevant asset events
have been produced.